// PIPREVENDA/frontend/src/types/api.generated.ts
export interface components {
    schemas: {
// ...
        ListarContratosResponse: {
            /** Format: int32 */
            id: number;
            contratoId: string;
            pcr: string;
            clienteNome: string;
            clienteCnpj: string;
            clienteNrSap: string | null;
            clienteUf: string | null;
            grupoEconomicoNome: string | null;
            segmento: components["schemas"]["Segmento"];
            tipo: components["schemas"]["TipoContrato"];
            /** Format: double */
            galonagemContratada: number;
            /** Format: double */
            galonagemFaturada: number;
            /** Format: double */
            galonagemPercentual: number;
            situacaoPcr: components["schemas"]["SituacaoPcr"];
            /** Format: date */
            fimVigencia: string;
            denuncia: boolean;
            bandeira: string;
            diretoria: string;
            regionalVendas: string;
            pontoVenda: string;
            consultor: string;
            papelGuardaChuva: components["schemas"]["NullableOfPapelGuardaChuva"];
        };
// ...
        PaginationResponseOfListarContratosResponse: {
            items: components["schemas"]["ListarContratosResponse"][];
            /** Format: int32 */
            pageIndex: number;
            /** Format: int32 */
            pageSize: number;
            /** Format: int32 */
            totalCount: number;
            /** Format: int32 */
            totalPages?: number;
        };
// ...
        Segmento: "Rede" | "Grr" | "Cofa" | "Cofd" | "Cofar" | "Outros" | "Spot";
        /** @enum {unknown} */
        SituacaoMes: "Ativo" | "Inativo";
        /** @enum {unknown} */
        SituacaoPcr: "Vigente" | "VencidoPorGalonagem" | "VencidoPorData";
        /** @enum {unknown} */
        TipoContrato: "Pcvm" | "Imagem" | "Comodato";
        /** @enum {unknown} */
        TipoEventoContrato: "NovoNegocio" | "Renovacao" | "Readequacao" | "Cessao" | "Sucessao" | "Denuncia" | "Encerramento";
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}